Change amount of loan needed to double column

+++ app/Http/Controllers/Api/Project/ProjectController.php
@@ -101,7 +101,7 @@ class ProjectController extends ApiController
             ->integer('minParticipations')->alias('min_participations')->next()
             ->boolean('isMembershipRequired')->alias('is_membership_required')->next()
             ->boolean('isParticipationTransferable')->alias('is_participation_transferable')->next()
-            ->integer('amountOfLoanNeeded')->alias('amount_of_loan_needed')->next()
+            ->double('amountOfLoanNeeded')->alias('amount_of_loan_needed')->next()
             ->get();
 
         $project = new Project();
@@ -156,7 +156,7 @@ class ProjectController extends ApiController
             ->integer('minParticipations')->alias('min_participations')->next()
             ->boolean('isMembershipRequired')->alias('is_membership_required')->next()
             ->boolean('isParticipationTransferable')->alias('is_participation_transferable')->next()
-            ->integer('amountOfLoanNeeded')->alias('amount_of_loan_needed')->next()
+            ->double('amountOfLoanNeeded')->alias('amount_of_loan_needed')->next()
             ->get();
 
         $project->fill($data);
